clippy auto-fix

+++ storage/benches/bench_mvcc_storage.rs
@@ -53,7 +53,7 @@ fn populate_storage(storage: &MVCCStorage, keyspace_id: KeyspaceId, key_count: u
     println!("Keys written: {}", key_count);
     let snapshot = storage.open_snapshot_read();
     let prefix: StorageKey<'_, 48> =
-        StorageKey::Reference(StorageKeyReference::new(keyspace_id, ByteReference::new(&[0 as u8])));
+        StorageKey::Reference(StorageKeyReference::new(keyspace_id, ByteReference::new(&[0_u8])));
     let iterator = snapshot.iterate_prefix(prefix);
     let count = iterator.collect_cloned_vec::<BUFFER_KEY_INLINE, BUFFER_VALUE_INLINE>().unwrap().len();
     println!("Keys confirmed to be written: {}", count);
